requirements.txt
# Core Dependencies
Flask==2.3.3
Werkzeug==2.3.7
Jinja2==3.1.2
itsdangerous==2.1.2
click==8.1.7

# Database
Flask-SQLAlchemy==3.1.1
SQLAlchemy==2.0.23
Flask-Migrate==4.0.5
psycopg2-binary==2.9.9  # For PostgreSQL

# Authentication & Security
Flask-Login==0.6.3
Flask-WTF==1.2.1
email-validator==2.1.0.post1
python-dotenv==1.0.0
Flask-Talisman==1.0.0
Flask-Limiter==3.5.0

# Data Processing
pandas==2.1.1
numpy==1.26.0
scikit-learn==1.3.2
joblib==1.3.2
python-dateutil==2.8.2
pytz==2023.3.post1

# Data Visualization
plotly==5.17.0
matplotlib==3.8.0

# API & Web
Flask-RESTful==0.3.10
Flask-Cors==4.0.0
Flask-Babel==3.1.0
Flask-Bootstrap==3.3.7.1
Flask-Mail==0.9.1
Flask-Admin==1.6.1

# Development & Testing
pytest==7.4.2
pytest-cov==4.1.0
coverage==7.3.2
black==23.9.1
flake8==6.1.0
isort==5.12.0

# Production
gunicorn==21.2.0
gevent==23.7.0

# Utilities
python-dotenv==1.0.0
requests==2.31.0
python-slugify==8.0.1
Pillow==10.0.1
babel==2.13.0
redis==5.0.1
